webnfc/layout/push.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/nfchelp.js"></script>
<script>
'use strict';
const invalid_messages =
    [
      // NFCMessage.data: should have at least 1 valid record.
      // https://w3c.github.io/web-nfc/#the-push-method - Step 8.
      createMessage([{}]),
      // https://w3c.github.io/web-nfc/#dfn-map-text-to-ndef
      // NFCRecord must have data.
      createMessage([createTextRecord()]),
      // NFCRecord.mediaType for 'text' record must be 'text/*'.
      createMessage([createRecord('text', 'application/json',
          test_number_data)]),
      // NFCRecord.data for 'text' record must be number or string.
      createMessage([createTextRecord(test_buffer_data)]),
      createMessage([createTextRecord(test_json_data)]),
      // https://w3c.github.io/web-nfc/#dfn-map-a-json-object-to-ndef
      // NFCRecord must have data.
      createMessage([createJsonRecord()]),
      // NFCRecord.data for 'json' record must be object.
      createMessage([createJsonRecord(test_buffer_data)]),
      createMessage([createJsonRecord(test_number_data)]),
      createMessage([createJsonRecord(test_text_data)]),
      // NFCRecord.mediaType for 'json' record must be 'application/json'.
      createMessage([createRecord('json', 'image/png', test_json_data)]),
      // https://w3c.github.io/web-nfc/#dfn-map-a-url-to-ndef
      // NFCRecord must have data.
      createMessage([createUrlRecord()]),
      // NFCRecord.data for 'url' record must be string.
      createMessage([createUrlRecord(test_buffer_data)]),
      createMessage([createUrlRecord(test_number_data)]),
      createMessage([createUrlRecord(test_json_data)]),
      // https://w3c.github.io/web-nfc/#dfn-map-binary-data-to-ndef
      // NFCRecord must have data.
      createMessage([createOpaqueRecord()]),
      // NFCRecord.data for 'opaque' record must be ArrayBuffer.
      createMessage([createOpaqueRecord(test_text_data)]),
      createMessage([createOpaqueRecord(test_number_data)]),
      createMessage([createOpaqueRecord(test_json_data)])
    ];

// promise_test(t => {
//   return promise_rejects(t, "TypeMismatchError", navigator.nfc.push(undefined));
// }), "Test that passing undefined to nfc.push would raise TypeMismatchError.";
// 
promise_test(t => {
  invalid_messages.forEach(message => {
    console.log(message);
    console.log('DEBUG-1');
        return promise_rejects(t, "SyntaxError", navigator.nfc.push(message));
  })
}, "Test that promise is rejected with SyntaxError if NFCMessage is invalid.")

invalid_messages.forEach(message => {
  promise_test(t => {
    return promise_rejects(t, "SyntaxError", navigator.nfc.push(message));
  }, "Test that promise is rejected with SyntaxError if NFCMessage is invalid : "+message[0]);
});
// 
// promise_test(t => {
//   return promise_rejects(t, "NotSupportedError", navigator.nfc.push(test_text_data));
// }, "nfc.push should fail when NFC HW is disabled.")
// 
// promise_test(t => {
//   return promise_rejects(t, "NotSupportedError", navigator.nfc.push(test_text_data));
// }, "nfc.push should fail when NFC HW is not supported.");
// 
// promise_test(t => {
//   return navigator.nfc.push(test_text_data, { timeout: 1 });
// }, "nfc.push should succeed when NFC HW is enabled.");
// 
// promise_test(t => {
//   return promise_rejects(t, "TimeoutError", navigator.nfc.push(test_text_data, { timeout: 1 }));
// }, "nfc.push should fail with TimeoutError when push operation is not completed before specified timeout value.");
// 
// //promise_test(t => {
// //  let message = createMessage([createTextRecord(test_text_data), 
// //                               createJsonRecord(test_json_data),
// //                               createOpaqueRecord(test_buffer_data),
// //                               createTextRecord(test_number_data),
// //                               createUrlRecord(test_url_data)],
// //                               test_message_origin);
// //  return navigator.nfc.push(message).then(() => {
// //    assert_equals(message, nfc.);
// //  })
// //});
// promise_test(t => {
//   return navigator.nfc.push(test_text_data).then(() => {
//     assertNFCPushOptionsEqual(createNFCPushOptions('any', Infinity, true), test_text_data);
//   })
// }, "Check that default NFCPushOptions values are correctly set.");
// 
// promise_test(t => {
//   let nfcPushOptions = createNFCPushOptions('tag', 1, false);
//   return navigator.nfc.push(test_text_data, nfcPushOptions).then((pushOptions) => {
//     assert_equals(nfcPushOptions, pushOptions);
//   })
// }, "Check that provided NFCPushOptions values are correctly converted.");
// 
// promise_test(t => {
//   return navigator.nfc.push(test_text_data)
//     .then(() => {
//       navigator.nfc.cancelPush;
//     });
// }, "nfc.cancelPush should succeed if there is pending push operation.");
// 
// promise_test(t => {
//   return navigator.nfc.push(test_text_data, { timeout :100 });
//   return promise_rejects(t, "AbortError", navigator.nfc.cancelPush());
// }, "nfc.cancelPush should reject pending promise with AbortError.");
// 
// promise_test(t => {
//   return promise_rejects(t, "NotSupportedError", navigator.nfc.push(new ArrayBuffer(32 * 1024 + 1)));
// }, "Reject promise with NotSupportedError if NFC message size exceeds 32KB.");

</script>
